Allow to use base_uri option for url routing.
Right now, the following methods are used to determine request_uri(),
used in routing :

* u or uri GET params
* PATH_INFO (index.php/uri)
* QUERY_STRING (index.php?/uri)
* REQUEST_URI, which is just the equivalent of the two above methods
  but for webservers which doesn't set PATH_INFO or QUERY_STRING

The problem with this is that none is controllable via user
configuration.

This patch modify the last method to try to use option('base_uri') in
addition to the internally computed base_path. For example, if
REQUEST_URI is /app/myroute and option('base_uri') is '/app', then
request_uri() (if not other methods are used) will return '/myroute'.

The main usage of this new method will be to allow rewrite rule of
this form :

 RewriteCond %{REQUEST_FILENAME} !-f
 RewriteRule ^(.*)$ index.php [QSA,L]

Which is the preferred method for "pretty"-urls in Nginx.